Adding the vis of points one has already visited to the map

+++ Streetly/Dataitem.swift
@@ -1,5 +1,5 @@
 //
-//  Dataitem.swift
+//  DataItem.swift
 //  Streetly
 //
 //  Created by Mojmír Horváth on 31.07.24.
@@ -9,20 +9,18 @@ import Foundation
 import SwiftData
 
 @Model
-class DataItem: Identifiable{
+class DataItem: Identifiable {
     var id: String
-    var lon: Float
-    var lat: Float
+    var lon: Double
+    var lat: Double
     var time: Date
     var cartype: String
-    
-    
-    init(lon: Float, lat: Float, car: String) {
+
+    init(lon: Double, lat: Double, car: String) {
         self.id = UUID().uuidString
         self.lon = lon
         self.lat = lat
         self.cartype = car
         self.time = Date()
     }
-    
 }
